backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
django-otp==1.2.0
django-otp-plugins==1.2.0
